Local storage: Restore task/project type on load
Since last commit, it was impossible to delete tasks and projects that
were retrieved from the localstorage. The deletion function uses the
instance type (TASK or PROJECT classes) on its logic, but this info
was not being preserved

Now, each entry on tasklist/projectlist is used to create a new object
of Task/Project class. It was necessary to change those classes'
constructor functions and how the "getLocalStoredTask/Project" functions
handle the task/project list